Fix scheduling bug of delayed vs. non-delayed messages.
There was indeed a bug (as I'd suspected) when proc P1 sends two messages
to proc P2, the first when a message is delayed and a second when the
message is not delayed.  The 2nd message arrives first, oops.

While fixing the bug, I removed the process state's distinction between the
'outbox' (outgoing non-delayed messages) and 'delayed' (outgoing delayed
messages).  All messages now have a delay, and bang() determines if the
delay should be 0 or greater-than-0.

Also, some variable renaming and fix run_scheduler_with_tokens_test() to
give that test more scheduler tokens to cope with the fact that scheduling
consumes slightly more tokens than it used to.