Make simulator compatible with QuickCheck Mini, and change simulator return tuple.
It turns out that the only part of the simulator (so far) that relies on
a commercial QuickCheck feature is the use of `conjunction`.  But that's
easy to replace, using an `andalso` clause.  I have not retrofitted all
of the example modules yet, but it's quite easy to do.

Also, I've given more flexibility to the simulator implementation
callbacks' return tuple: the function can specify an atom instead of
a fun, e.g. `{recv_general, next_callback, State}` instead of
{recv_general, fun next_callback/2, State}`.  If you use the atom
flavor, then the atom must be the same as a function (of arity 2)
that exists in the simulator module.  If you wish to specify a
function outside of the simulator module, then you must use the
fun flavor of the tuple.  (Also, don't name your function `same`,
since that atom has special meaning within the simulator!)