Scale back my plans for the 2-phase protocol: eliminate server's 'waiting'.
My initial idea for this distributed protocol, to try to make it from
something unfair to something (at least somewhat) fair, a server would
act in a manner like this when two clients attempt a set op simultaneously:

    C1                          C2                          Server
    |-------- phase 1 ask ----------------------------------->
                                |-------- phase 1 ask ------->
    <-------- phase 1 ok + cookie + current value V0 --------|
                                <-------- phase 1 sorry -----|
    |-------- phase 2 set + cookie + new value V1 ----------->
    <-------- phase 2 ok ------------------------------------|
                                <- phase 1 ok+cookie+val V1 -|
                                |- phase 2 set+cookie+val V2->
                                <- phase 2 ok ---------------|

So, the loser of phase 1, C2, would receive receive 2 messages in
response to its phase 1 ask:
   1. phase 1 sorry
   2. phase 1 ok + cookie + current value

My reasoning behind this is that the "ask" would be like asking for a
place to wait in a line/queue: the server would maintain this ordered
queue and return "phase 1 ok..." messages to each client in the order
that they were received.

I have a sneaking suspicion that this scheme is subject to nasty,
subtle races.  So I am removing the whole idea for now.  I might try
adding it later, if I can convince myself that my deadlock worries
are misplaced.