Add ph1_cancel operation to help with multiple clients.  It's buggy.  :-)
So, the bug in this one is a bit insidious.  In this counterexample,
client c1 ends up with a quorum failure at step #25:

    {c1,25,{ph1_quorum_failure,c1,num_oks,1}}

So c1 sends a ph1_cancel to the 1 server that it did get an ok from, s1.
However, that message is dropped.  At time 34, c1 resends the ph1_cancel
message to s1.  However, at time 36, the scheduler sends s1 a timeout,
which means that server s1 moves from server_asked -> server_unasked state,
so when the 2nd ph1_cancel message is arrives, that message is ignored
by the s1.  The client then goes into an infinite resend loop, and
server s1 never sends a reply.  Sweet.