Attempt to add vclock-only control logic over strict counters ... fails.
So, I should know this, but I didn't *know* it in my innermost brain.
Vector clocks, by themselves, cannot solve my distributed counter problem
without also having app-specific resolution logic.

The distrib_counter_2phase_sim.erl simulator uses a very simple,
app-specific method for figuring out the next value of the counter:

    cl_p1_send_do(C = #c{clop = ClOp, ph1_oks = Oks}) ->
        [{_, _, _, _, MaxVal}|_] = lists:reverse(lists:keysort(5, Oks)),
        NewVal = MaxVal + 1,
        ...

I had this great idea that I'd use vclocks to make this decision instead.
Why?  Because the next step of the evolution of this protocol is to add an
unconditional-set to the protocol.  Then the general outline of the
protocol would look like:

    1. Phase 1: send ask ops to all servers
    2. If we do not get a quorum of oks, then stop.
    3. Phase 2: send unconditional-set ops to all servers, not ok only servers.

The problem with this is that the unconditional-set could be delayed and
therefore clobber a newer value set by someone else.  That shouldn't
violate the correctness of the protocol(*), but it could clobber values on
a minority of servers, which combined with a server failure later, might
cause problems?

(*) I.e. screw up a quorum majority of replicas.

For example:

    0. We have clients c1 & c2 and servers s1, s2, & s3.
       All servers have counter = 0.
    1. c1 gets ask oks from s1 & s2.
    2. c1 sends unconditional-sets of counter=1 to s1 & s2.
       - s1's message arrives promptly, s2's is delayed.
    3. c2 gets ask oks from s1 and s3; s2 says sorry.
    4. c2 sends unconditional-sets of counter=2 to s1 & s3.
    5. c1's message to s2 arrives at s2.
    6. Current view of things:
         s1 counter=2
         s2 counter=1
         s3 counter=2

Hm, well, that isn't a big problem.  But perhaps there's a bigger problem
lurking in there?

Well, back to the current checkin.

The problem is that vclocks alone aren't enough.  Suppose this situation:

    0. We have clients c1 & c2, and servers s1, s2, & s3.
       All servers have counter = 0.
    --
    1. c1 gets ask oks from s1, & s2: all agree that vc=[] and counter=0
    2. c1 sends phase2 sets of vc=[{c1,1}], counter=1 to all ok servers.
       - s1's message arrives promptly, s2's is dropped.
    --
    3. c2 gets ask oks from s2, & s3: all agree that vc=[] and counter=0
    4. c2 sends phase2 sets of vc=[{c2,1}], counter=1 to all ok servers.
    --
    5. c2 gets ask oks from s2, & s3: all agree that vc=[{c2,1}] and counter=1
    6. c2 sends phase2 sets of vc=[{c2,2}], counter=2 to all ok servers.
    --
    7. Snapshot of server state: s1: vc=[{c1,1}], counter=1
                                 s2: vc=[{c2,2}], counter=2
                                 s3: vc=[{c2,2}], counter=2
    8. c1 gets ask oks from s1 & s2, but we have sibling counter values.
       The vclocks told us that there's a conflict, but the vclocks don't
       have enough info to tell us if the counter's "true" value is 1 or 2.

Duh.  Bryan O'Sullivan said it pretty well at Erlang Factory San Francisco
2011: (paraphrasing) "Vector clocks are like a fire alarm: they will tell
you that there is a fire, but they cannot put the fire out."