Add distrib_counter_2phase_vclockset_sim.erl, which adds an unconditional-set operation to the same protocol simulated by distrib_counter_2phase_vclock_sim.erl
This unconditional set operation is used in phase 2 of the protocol,
sent to all servers that did not respond 'ok' to the phase 1 ask operation
or did not respond at all.  The purpose for this operation is to try to
reduce the amount of work that a read repair-style operation would require
to fix out-of-sync servers.

The vclock is used to handle cases where an unconditional set arrives
after a new operation has mutated the value again: the vclock of
the object in the unconditional set op should be an ancestor and thus the
late-arriving op would be ignored.

Also, new QuickCheck stats are added to measure:

    * The number of unconditional set ops sent
    * The number of unconditional set ops delayed
    * The number of unconditional set ops that were ignored (the server's
      original value is used)
    * The number of unconditional set ops that were accepted (the op's
      new value is used)
    * The number of unconditional set ops that resulted in something else.
      Right now, only the counter is included in the contents, so
      reconciling the sibilings is easy.