New sim module with major protocol extension: change notification.
distrib_counter_2phase_vclocksetwatch_sim.erl is based on the module
distrib_counter_2phase_vclockset_sim.erl, for comparing by "diff" or
by eye.

See the "Protocol implementation" section for a small sketch of the
change notification protocol.  For a complete MSC diagram, see the
doc/msc/notify_proto.* files.

This implementation is likely buggy.  However, it manages to pass 10K
test cases for 2 clients and 1 server:

    eqc:quickcheck(eqc:numtests(10*1000,slf_msgsim_qc:prop_simulate(distrib_counter_2phase_vclocksetwatch_sim, [{max_clients,2}, {max_servers,1}]))).

Note that there are no verification tests in the verify_property()
function for checking the correctness of the change notification trace
events.  Look for the several "TODO" and "todo" strings throughout
the source.