Fix a really cool deadlock bug; interesting that I didn't see it earlier.
So, this fixes a deadlock bug where a client is in client_ph1_cancelling
state, and a server is in server_change_notif_toindividuals state.
Both states are requiring a synchronous reply from the other, bummer.

This creates an interesting TODO question: is it necessary for
for _all_ of the client counter-related states to explicitly handle
the watch_notify*_req messages to avoid deadlock?