Add {stop_step, N} option to simulator options.
If `{stop_step, N}` is in the `Options` list passed to
`slf_msgsim:new_sim()`, then the simulator will stop execution at that
simulated execution step.  If a simulation stops early, then the odds are
almost certain that your `verify_property()` function will fail, so don't
be surprised when if fails.