Add a new state to the client watcher op: client_watch_waiting_2 (with slight "bug"!)
Remember, way back in the Git commit log, that I'd said that the
"watch maybes" statistic was always zero?  That isn't a good thing.
So, I figured out why.

A watcher client sees a `{watch_notify_maybe_req,...}` message
because earlier attempts to deliver a definite `{watch_notify_req,...}`
message have failed.  That failure requires at least one timeout
event to happen.

However, the way the message simulator works, a timeout event cannot
be delivered until all processes in `recv_timeout` mode are blocked
trying to receive a message.  That *includes* the watch client itself,
waiting for a definite or maybe notification.  When that all-procs-
are-blocked situation happens, *all* procs will get a timeout event,
including the watch client.  The watch client then moves on with its
life, and when the maybe notification arrives (sometime much later),
that proc no longer cares.

So, as a hack for the simulator, I've added a new state to the watch
client: `client_watch_waiting_2`.  When a client is in
`client_watch_waiting` state and gets a timeout, it moves to
`client_watch_waiting_2` state.  If a timeout has also simultaneously
delivered to a counter client and/or a counter server, that proc(s)
now has the opportunity to send a `{watch_notify_req,...}` or
`{watch_notify_maybe_req,...}` message and have a reasonable chance
that the watch client will still be in `client_watch_waiting_2` state.

However, there is a slight bug in this, as detected by the event
trace check C4.  That will be discussed and fixed in the next commit.