Fix C4 verification test in distrib_counter_2phase_vclocksetwatch_sim.erl.
An example of the "problem" looks like this:

1. Watch client W1 successfully sets up a watch.
2. Counter client C1 tries to update a counter but fails in phase 2.
3. One of the servers involved in step #2 attempts to send a
   maybe notification to W1.  The attempts fail due to network partition.
   The server keeps retrying....
4. Counter client C1 tries to update the counter again.  This time it
   succeeds.
5. The notification described in step #3 is finally delivered to W1.

The C4 check inside distrib_counter_2phase_vclocksetwatch_sim.erl gets
confused by this, because the `{ph2_do_set,...}` messages sent by both
attempts (step #2 and step #4) appear to be the same: the counter value
is the same, and the vector clock is the same (if both attempts happen
within the same wall clock second, which is usually true).  Therefore,
the C4 test can't tell them apart and thinks that the notification
described in step #5 came from the update from step #4 (which is not
true).

In order to make it easy for C4, we help make the vector clocks for
the two counter updates unique by using `erlang:now()` instead of
the usual `calendar:datetime_to_gregorian_seconds(erlang:universaltime())`.