1st attempt at getting a McErlang simulation ready: distrib_counter_bad1_sim
This doesn't do anything directly with McErlang yet.  But it's
getting closer, I hope.  I want/need something that can reuse as
much of the msgdropsim infrastructure as possible ... but I haven't
spent enough time to figure out the easiest/laziest/most beautiful
way to do that yet.  As a result, this commit is pretty ugly.

Added a new QuickCheck property creator-thingie to slf_msgsim_qc.erl,
`prop_mc_simulate/2`.  It usage is indended to be 100% compatible
with `prop_simulate/2`.  Right now it's only an ugly kludge, probably
only good enough for `distrib_counter_bad1_sim.erl`, but it's a
start.

Example:

    eqc:quickcheck(eqc:numtests(5000, slf_msgsim_qc:prop_mc_simulate(distrib_counter_bad1_sim, []))).

For `distrib_counter_bad1_sim.erl` there have been a lot of changes,
many of them temporary kludges:

* Modify `general_initial_ops/4` to be able to take a list of
server pids out of the `Props` property list.  The kludge is that
the server processes need to be started first so that their PIDs
can be embedded into the client op tuple because we aren't using
registered names so we gotta use PIDs, bleh.  I may yet go back
to using registered names, I think, it may make life a heckuvalot
easier ... except for the race conditions that may pop up when
starting all of these silly processes, bleh.

* Create a `verify_mc_property/7` function to perform the same
test sanity check as `verify_property/7` does for msgdropsim's
simulated processes.

* Convert the entire process simulation code to standard Erlang
code.  I'd *love* to find a better way to eliminate this step
or to automate the code generation from this step, because it's
ugly and manual (and therefore error-prone), but the wonderful-
in-the-real-world-but-difficult-to-model semantics of selective
receive make that job difficult.  So, I need to put a lot more
thought into this aspect, alas.

For each state, e.g., `counter_client1_reply()`, there is an
Erlang function that corresponds to that FSM state, e.g.,
`e_counter_client1_reply()`.  The `e_` prefix stands for
"Erlang".