Hrm, it looks like 2 receive statements "nested" don't work.
If Proc A is waiting for a message from Proc B with timeout 5X,
and Proc B is waiting for a message with timeout 1X, then McErlang
can decide that Proc A's timeout can happen first, even when

So, we'll avoid the problem by removing an "after" clause when
getting results from clients.  As a result, if a client is misbehaved
and doesn't send us the results message, then we'll hang forever.