Intermediate checkin for McErlang tests: nothing fundamentally new, alas.
So, I'm trying to figure out why the
distrib_counter_2phase_vclocksetwatch_sim algorithm blows up
so incredibly badly with McErlang.  1 client + 3 servers + that
single client issuing a single `counter_op` = 22 *million* states.
That's insane.  Something is wrong.  Or, at least I hope so.  The
alternative is that I'm simply not used to dealing with this kind
of exponential state explosion.  {sigh}  The challenge/trick: how
to figure out the difference!